Add schema for handling payments
This is just a bit of exploration, it will likely need editing when we
figure out how to do things properly. Some open questions:

 * It would be nice to have a custom token for every subscription, and
   have a contract mint that, then you can buy subscriptions and give
   them away as a gift. It would also enable login without an account;
   you can prove that you own the subscription tokens, and that should
   be sufficient. Is this feasible?

 * If it can't be done with a contract, or if it's tedious, should we
   mint the token using cardano-cli?

 * Before we have payment requests, shouldn't we have subscriptions to
   buy?

We can worry about all of this later though. For now, this is a bit of
schema to experiment with. The next steps will be to generate keys to
receive commission, and to watch for transactions that fill the payment.