Begin working on cardano-wallet client
So cardano-wallet happens to be written in Haskell and it uses Servant
for its API ... which means we could get a client for free. And indeed,
the cardano-wallet CLI app uses a Servant client internally.

Unfortunately, none of the Cardano packages are published to Hackage,
they all depend on each other, and they all use this bespoke build
system that is incompatible with the regular Haskell ecosystem.

I tried briefly if I could just copy only the files we need, and maybe
it would compile, but the dependency tree is vast, and I stumbled upon a
bech32 dependency which is marked broken in Nixpkgs. I could try to get
all of this working, but honestly I think I would spend an evening on it
and end up very frustrated and unhappy without much progress. So instead
I'll write my own client. (Cue Bender quote here.)

As a first step, let's see if we can query the network info. A first
step for that is to define deserializers.