Ensure test env overrides original env for database
The tests were failing for me locally, because it used the PG* variables
from the original environment instead of the ones that the test
provides, presumably because we prepend them, and the later ones take
precedence? To avoid any ordering issues, put things in a HashMap to
ensure only the new values end up in the environment, not the old ones.